requirements.txt
pykeepass 
mega.py 
python-dotenv
pycryptodome
